requirements.txt
PyQt5==5.15.10
python-vlc==3.0.20123
